Fix build by adjusting ring color syntax
