+++ src/index.css
@@ -68,7 +68,7 @@ a:focus-visible,
 input:focus-visible,
 textarea:focus-visible,
 select:focus-visible {
-  @apply outline-none ring-2 ring-offset-2 ring-primary-600/50;
+  @apply outline-none ring-2 ring-offset-2 ring-primary-600 ring-opacity-50;
 }
 
 ul[role='list'] {
